new approach for detecting falling floors
Old approach was reliant on using the floor's rigidBody.velocity to determine if the block was falling. Turns out the walls don't consistently have rigidBodies, so this didn't always work. New approach simply compares the position of the block against its position a frame ago and should be more generally applicable