VictoryPanelVisibility
I just changed the house prefab to have the victory panel set to inactive by default. It will no longer block the view in the game window in the editor but will have no issue becoming active when victory is achieved.

Also made the extremely minor change of linking the "Time UI" field for the Timer script in the Template Level in case anyone needs to see how it should be set