Fix Fox_Animator_Controller
Looks like at some point in restoring a piece of the animation controller got lost. This is just re-adding the right transition/animation nodes back to the controller. The scripts were correct, and throwing errors because of trying to access the missing components, so this just adds back what they need!