Undo LETS GOOOOO
Added an undo manager and hooked it up to the turn-based-character script(s). Keeps a stack of turn states (everything that moved in a turn) and can pop off turns to undo them.
Also changed the destroyBounds script to hide the game objects instead of destroying them because otherwise you can't move them back.
Added the new Undo Manager script to the GameManager prefab.