Level-Complete-Condition Bug Fix
Fixed the bug where the game wouldn't arrest player control when the fox walked into a house only in a built version of the game but worked fine in editor.

For anyone curious: issue had to do with script execution order.
Changed it to some initialization stuff that needed to happen is moved to Awake() instead of Start() since Awake is called first. That fixed the issue.