Polishing
turnManager.say() now automatically determines the height of the text box based on the message length. Some messages "Well done!, Good job!, etc." use a less-wide text box.

Made a new prefab called "MessageTriggers" that can be used to have Summer say a message when the trigger is tripped.

Added the cool fairy material to Summer in all cases.

Slightly changed the font to have a better comma "," and exclamation point "!"

Added a bunch of message triggers to Level_2 to tell the player how to undo a move if they mess something up. Formally introduced in level_3