Update Leon_Level_17.unity
Some quick minor adjustments to help fix confusion testers had. Rotated the house to face an appropriate direction so testers don't think that they cant get to the house because the door is facing a different way. I also added a couple rocks on top of cliffs so the testers do not think that that is the correct way to go to solve the puzzle.