Update Leon_Level_25.unity
I cut back the cliffs on the level so I could add grass to help guide the player more. Rounded it out and cut back as much empty space as I could to keep the round shape. I left the puzzle intact because I thought it was a good puzzle.